requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
qrcode
python-multipart
jinja2
pillow
itsdangerous
httpx